+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [Unreleased]
+- Add `EnvelopeRecipient.intermediaryName` and `EnvelopeRecipient.intermediaryEmail`
+
 ## [1.3.0] - 2021-09-03
 ### Added
 - Add `EnvelopeTemplateDocument.positions` endpoint